requirements.txt
Django>=2.1.0,<2.2.0
graphene-django==2.2
graphene-mongo==0.2.0
django-filter==2.0.0
django-graphql-jwt==0.1.5
mongoengine>0.15.3,<=0.16.3
pymongo>3.7.0,<=3.7.2

flake8>=3.6.0,<3.7.0